Address GLM-DSA package review feedback
